requirement.txt
Pillow==9.2.0
Flask==3.0.3
google-api-python-client
matplotlib==3.6.2
numpy==1.25.0
pandas==1.4.3
requests==2.31.0
scikit-learn==1.3.2
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.1
Werkzeug==3.0.1
